% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileData.R
\name{compileData}
\alias{compileData}
\title{compileData}
\usage{
compileData(
  detection.folder = "Detections",
  file.ext = ".csv",
  sep.type = ",",
  save = TRUE,
  remove.duplicates = T,
  save.duplicates = F,
  split = T
)
}
\arguments{
\item{detection.folder}{The name of the folder containing the detection data}

\item{file.ext}{The extension of your detection files (e.g. ".csv")}

\item{sep.type}{The column separator in your detection files (e.g. ",")}

\item{save}{Should the compiled data be saved in the detections folder (TRUE or FALSE)
if FALSE the compiled detections will be in your R environment only}

\item{remove.duplicates}{Should the duplicated data (i.e. detection of the
same tag, on the same receiver at the same time) be removed (TRUE or FALSE)}

\item{save.duplicates}{Should the duplicated detections be saved in your detections folder (TRUE or FALSE)
if FALSE the  duplicated detections will be in your R environment only}

\item{split}{Should the data be processed in chunks (TRUE or FALSE; recommended). Lowers the chances of
errors due to the memory limit being reached when the data are too numerous}
}
\value{
A data frame object that contains your compiled detections, and another that contains the duplicates
}
\description{
Compiles the acoustic telemetry data in a folder into a file that is useable by ATfiltR.
}
\examples{
\dontrun{
# Before running the function, you must be in an R project that contains your detections folder
function(detection.folder="Detections", file.ext=".csv",
sep.type=",", save=TRUE, remove.duplicates=T, save.duplicates=F)
}
}
