% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dependencies.R
\name{loadRData}
\alias{loadRData}
\title{loadRData}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{the file you are loading}
}
\value{
Load an RData file into a specified variable name
}
\description{
loadRData
}
