% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSolo.R
\name{loadSolo}
\alias{loadSolo}
\title{loadSolo}
\usage{
loadSolo(detection.folder = "Detections")
}
\arguments{
\item{detection.folder}{The name of the folder containing the detection data}
}
\value{
The data loaded for findSolo()
}
\description{
Loads the data for use in findSolo()
}
\examples{
\dontrun{
# Before running the function, you must be in an R project that contains
your detections folder and data folder, your data must be compiled via
compileData(), and could (should) have already been filtered by wWindow()
findSolo(detection.folder="Detections", save.solo=T, save=T, delay = 1)
}
}
