% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSpeed.R
\name{loadSpeed}
\alias{loadSpeed}
\title{loadSpeed}
\usage{
loadSpeed(detection.folder = "Detections")
}
\arguments{
\item{detection.folder}{The name of the folder containing the detection data}
}
\value{
The data loaded for speedCheck()
}
\description{
Loads the data for use in speedCheck()
}
\examples{
\dontrun{
# Before running the function, you must be in an R project that contains
your detections folder and data folder, your data must be compiled via
compileData(), and have been filtered by wWindow(). It should (but doesnt have to) also have gone through findSolo()
findSolo(detection.folder="Detections", save.solo=T, save=T, delay = 1)
}
}
