% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dependencies.R
\name{myView}
\alias{myView}
\title{myView}
\usage{
myView(x, title)
}
\arguments{
\item{x}{What you want to View}

\item{title}{the title of what you want to View}
}
\value{
A View of a dataframe
}
\description{
myView
}
