% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toActel.R
\name{toActel}
\alias{toActel}
\title{toActel}
\usage{
toActel(
  detection.folder = "Detections",
  data.folder = "Data",
  target.folder = "Actel"
)
}
\arguments{
\item{detection.folder}{The name of the folder containing the detection data}

\item{data.folder}{The name of the folder containing auxillary data (e.g. spatial data). Can be se same af detection.folder}

\item{target.folder}{the name of the folder in which to store the data for use in actel}
}
\value{
Your detection, animal, spatial and deployment data prepared in the actel format and saved in your target folder
}
\description{
Converts your ATFiltR files for use into actel
}
\examples{
\dontrun{
# Before running the function, you must be in an R project that contains
your detections folder and data folder.
# You must have processed your data with at least compileData() and
withinWindow()
}

}
