% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wWindow.R
\name{wWindow}
\alias{wWindow}
\title{wWindow}
\usage{
wWindow(
  detection.folder = "Detections",
  data.folder = "Data",
  sep.type = ",",
  save.out.of.deployment = F,
  save.unknown.tags = T,
  discard.first = 24,
  save = T
)
}
\arguments{
\item{detection.folder}{The name of the folder containing the detection data}

\item{data.folder}{The name of the folder containing auxillary data (e.g. spatial data). Can be se same af detection.folder}

\item{sep.type}{The column separator in your data (e.g. ","). All dataframes that you plan on using should have the same separator}

\item{save.out.of.deployment}{Should the data collected outside of the deployment window be saved in the detections folder? (TRUE or FALSE)
if FALSE the compiled detections will be in your R environment only}

\item{save.unknown.tags}{Should the data from tags that aren't in your animal data (and from your own tags but outside of the deployment window) be saved in the detections folder? (TRUE or FALSE)}

\item{discard.first}{How many of the first hours after deploying a tag should be discarded? (e.g. 24 = the first 24h will be discarded) (if save.unknown.tags is TRUE the discarded data will be saved in the unknown tags file)}

\item{save}{Should the data be saved in your detections folder (TRUE or FALSE)
if FALSE the filtered detections will be in your R environment only}
}
\value{
A data frame object that contains the detections from your tags, collected during a deployment window, another that contains the detections outside of receiver deployment, and a third containing detections from unknown tags or tags pinging outside of their deployment window
}
\description{
Finds and filters out the detections that are outside of the deployment window (deployment of tags and/orr of receivers).
}
\examples{
\dontrun{
# Before running the function, you must be in an R project that contains your detections folder and data folder
wWindow(detection.folder="Detections", data.folder="Data", sep.type=",", save.out.of.deployment=F,
save.unknown.tags=T, discard.first=24, save=T)
}
}
