% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSolo.R
\name{findSolo}
\alias{findSolo}
\title{findSolo}
\usage{
findSolo(
  detection.folder = "Detections",
  save.solo = T,
  save = T,
  per.receiver = T,
  delay = 0.5
)
}
\arguments{
\item{detection.folder}{The name of the folder containing the detection data}

\item{save.solo}{Should the solitary detections be saved in the detections folder? (TRUE or FALSE)}

\item{save}{Should the data be saved in your detections folder (TRUE or FALSE)
if FALSE the filtered detections will be in your R environment only}

\item{per.receiver}{If TRUE (default) the solitary detections are considered solitary when they occur alone on a given receiver (Similarly to Kessel et al. 2014). If FALSE they are considered solitary
when they occur alone across the whole array.}

\item{delay}{(hours) defines solitary detections: detections that are recorded *delay* hours after the previous one and *delay* hours before the subsequent one are considered solitary.}
}
\value{
A data frame object that contains the detections from your tags, wihtout solitary detections, another that contains the solitary detections
}
\description{
Finds and filters out the detections that are alone within a defined time window (delay).
}
\examples{
\dontrun{
# Before running the function, you must be in an R project that contains your detections folder and data folder, your data must be compiled via compileData(), and could (should) have already been filtered by wWindow()
findSolo(detection.folder="Detections", save.solo=T, save=T, delay = 1)
}
}
